backend/requirements.txt
pillow
numpy
uvicorn
fastapi
python-multipart